extract question page handlers
